Graph class rework
The previous graph class was too heavy, so this is a rework of it. Instead of a graph class holding nodes and adding edges,
the vertices are assumed to be the ints 0-(n-1), and the edges are in a hashmap of hashmaps of weights of templatted type T
that the user directly interacts with.

Warning: when adding edges make sure to have a duplicate edge guard! Sometimes kattis problems will give the same edge twice with
different weights, and expect you to use the smaller one (only) for a shortest path.

This commit addresses bfs,dfs,shortest paths, scc, visits outside (a sink check), and min_heap. Not yet updated are MST algs and hopcroft karp.